twitter-dashboard/requirements.txt
twikit==2.1.3
python-dotenv==1.0.1
streamlit==1.39.0